Bugfix (sort of)
The class declares a variable $contents (ends in "S"), and in all other places, it's referred to as $this->content (without "S"). Since the retrieval is file_get_contents, I decided to rename all of them to contents-with-an-S.

This is not a bug per se, since variables set ($this->var="x") create them as needed. However, it seems to be not as intended.