Cleanup and rework HTTPProxyTFTPServer a bit
Change from $ to % for template, seams more standard
and does not break with bash.
Rename _uri to _url_template.
Remove unused debug constructor argument.
More unused rw cleanup.
Remove reduntant TFTPServer methods and fallback to default
implementations instead.
Fetch URL in get instead of exists, the only difference
is that on error TFTPServer will return error message
"Fail to read" instead of "File not found" but in both
cases the error code will be file not found error code.
If this is a problem i think we should change so that
return false from get is file not found also or
make it possible for the implementation to specfiy errors.